+++ picopt/handlers/container/archive/zip.py
@@ -4,8 +4,10 @@ from io import BytesIO
 from pathlib import Path
 from zipfile import ZIP_DEFLATED, ZIP_STORED, ZipFile, is_zipfile
 
+from typing_extensions import override
+
 from picopt.formats import FileFormat
-from picopt.handlers.container.archive import PackingArchiveHandler
+from picopt.handlers.container.archive import ArchiveClassType, PackingArchiveHandler
 from picopt.handlers.handler import INTERNAL
 
 
@@ -13,35 +15,42 @@ class Zip(PackingArchiveHandler):
     """Ziplike container."""
 
     OUTPUT_FORMAT_STR: str = "ZIP"
+    SUFFIXES: tuple[str, ...] = (".zip",)
     OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, archive=True)
     INPUT_FILE_FORMAT = OUTPUT_FILE_FORMAT
     INPUT_FILE_FORMATS = frozenset({INPUT_FILE_FORMAT})
     PROGRAMS = ((INTERNAL,),)
-    ARCHIVE_CLASS = ZipFile
+    ARCHIVE_CLASS: ArchiveClassType = ZipFile
 
     def __init__(self, *args, **kwargs):
         """Init delete_filenames."""
         super().__init__(*args, **kwargs)
 
+    @override
     @classmethod
     def _is_archive(cls, path: Path | BytesIO) -> bool:
         return is_zipfile(path)
 
+    @override
     @staticmethod
     def _archive_infolist(archive):
         return archive.infolist()
 
+    @override
     def _archive_readfile(self, archive, archiveinfo):
         return archive.read(archiveinfo.filename)
 
+    @override
     def _set_comment(self, archive) -> None:
         """Set the comment from the archive."""
         if archive.comment:
             self.comment = archive.comment
 
+    @override
     def _archive_for_write(self, output_buffer: BytesIO) -> ZipFile:
         return ZipFile(output_buffer, "x", compression=ZIP_DEFLATED, compresslevel=9)
 
+    @override
     def _pack_info_one_file(self, archive, path_info):
         """Add one file to the new archive."""
         zipinfo = path_info.archiveinfo.to_zipinfo()
@@ -57,6 +66,7 @@ class Cbz(Zip):
     """CBZ Container."""
 
     OUTPUT_FORMAT_STR: str = "CBZ"
+    SUFFIXES: tuple[str, ...] = (".cbz",)
     INPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, archive=True)
     INPUT_FILE_FORMATS = frozenset({INPUT_FILE_FORMAT})
     OUTPUT_FILE_FORMAT = INPUT_FILE_FORMAT
@@ -66,8 +76,9 @@ class EPub(Zip):
     """Epub Container."""
 
     OUTPUT_FORMAT_STR: str = "EPUB"
+    SUFFIXES: tuple[str, ...] = (".epub",)
     OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, archive=True)
     INPUT_FILE_FORMAT = OUTPUT_FILE_FORMAT
     INPUT_FILE_FORMATS = frozenset({INPUT_FILE_FORMAT})
     # never convert inside epubs, breaks src links.
-    CONVERT_CHILDREN = False
+    CONVERT_CHILDREN: bool = False