+++ picopt/handlers/image/jpeg.py
@@ -18,10 +18,10 @@ class Jpeg(ImageHandler):
     """JPEG format class."""
 
     OUTPUT_FORMAT_STR = str(JpegImageFile.format)
-    SUFFIXES = (".jpg", ".jpeg", ".mpo")
+    SUFFIXES: tuple[str, ...] = (".jpg", ".jpeg", ".mpo")
     OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, lossless=False, animated=False)
     INPUT_FILE_FORMATS = frozenset({OUTPUT_FILE_FORMAT})
-    PROGRAMS = (("mozjpeg", "jpegtran", "pil2jpeg"),)
+    PROGRAMS: tuple[tuple[str, ...], ...] = (("mozjpeg", "jpegtran", "pil2jpeg"),)
     _JPEGTRAN_ARGS_PREFIX = ("-optimize", "-progressive")
     # PIL Cannot save jpegs losslessly