+++ picopt/handlers/image/svg.py
@@ -5,18 +5,21 @@ from typing import BinaryIO
 
 from picopt.formats import SVG_FORMAT_STR, FileFormat
 from picopt.handlers.image import ImageHandler
-from picopt.handlers.non_pil import NonPILIdentifier
+from picopt.handlers.mixins import NonPILIdentifierMixin
 
 
-class Svg(ImageHandler, NonPILIdentifier):
+class Svg(ImageHandler, NonPILIdentifierMixin):
     """SVG format class."""
 
-    OUTPUT_FORMAT_STR = SVG_FORMAT_STR
-    OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, lossless=True, animated=False)
-    INPUT_FILE_FORMAT = OUTPUT_FILE_FORMAT
-    INPUT_FILE_FORMATS = frozenset({INPUT_FILE_FORMAT})
-    PROGRAMS = (("svgo", "npx_svgo"),)
-    _SVGO_ARGS = ("--multipass", "--output", "-", "--input", "-")
+    OUTPUT_FORMAT_STR: str = SVG_FORMAT_STR
+    SUFFIXES: tuple[str, ...] = (".svg",)
+    OUTPUT_FILE_FORMAT: FileFormat = FileFormat(
+        OUTPUT_FORMAT_STR, lossless=True, animated=False
+    )
+    INPUT_FILE_FORMAT: FileFormat = OUTPUT_FILE_FORMAT
+    INPUT_FILE_FORMATS: frozenset[FileFormat] = frozenset({INPUT_FILE_FORMAT})
+    PROGRAMS: tuple[tuple[str, ...], ...] = (("svgo", "npx_svgo"),)
+    _SVGO_ARGS: tuple[str, ...] = ("--multipass", "--output", "-", "--input", "-")
 
     def _svgo(self, exec_args: tuple[str, ...], input_buffer: BinaryIO) -> BytesIO:
         """Optimize using svgo."""