+++ NEWS.md
@@ -1,6 +1,15 @@
 # 📰 Picopt News
 
-## v5.0.0-a0
+## v5.1.0
+
+- gif2webp & img2webp experimental options hidden behind environment variables.
+  See README.
+
+## v5.0.1
+
+- Fix crash converting animated images with bad duration metadata.
+
+## v5.0.0
 
 - Alert!
   - Timestamps options changed, invalidating old timestamps. On your first run